feat: transform to aws lambda
