+++ .env.example
@@ -1,2 +1,3 @@
 SOURCE_IMAGE=
 LOCATION=
+MONGODB_ATLAS_CLUSTER_URI=
