--- .eslintignore
@@ -1,2 +1,0 @@
-build
-node_modules
