+++ .gitignore
@@ -1,3 +1,2 @@
 .env
-build
 node_modules
