--- tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": false,
-    "module": "commonjs",
-    "target": "es6",
-    "lib": ["es6"],
-    "moduleResolution": "node",
-    "removeComments": true,
-    "outDir": "./build",
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "resolveJsonModule": true
-  },
-  "include": [
-    "src/*",
-  ],
-  "exclude": ["node_modules"]
-}
